tests/dualipa/tests/requirements-test.txt
# Testing framework
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
pytest-asyncio>=0.21.0

# Core dependencies
torch>=2.0.0
transformers>=4.34.0
peft>=0.5.0
datasets>=2.14.0
unsloth>=0.3.1

# HTTP and async
aiohttp>=3.8.4
requests>=2.28.0

# Git operations
gitpython>=3.1.30
pygithub>=1.58.0

# Markdown parsing
markdown-it-py>=2.1.0
mistune>=2.0.4

# LLM integration
litellm>=0.1.685
openai>=0.27.8
tenacity>=8.2.0
pydantic>=1.10.8
cachetools>=5.3.0

# Utilities
loguru>=0.7.0 